[merge] Remove the merge package
This package is not very useful. It does not do anything special. It was created solely to present how to write adapters merging context values with logger entries. But it turns out that another package complicates things for newcomers. It will be better to remove it, but leave the example presenting how to merge context value with entry.