[contextadapter] Remove package contextadapter
This package is not very useful. It does not do anything special. It was created solely to present how to use other loggers bound to context.Context. But it turns out that another package complicates things for newcomers. It will be better to remove it, but leave the example how to use zap logger passed in the context.Context