[bugfix] Fix race condition in logger.Global
There is a chance that using two goroutines simultaneously, on an uninitialized global logger may lead to a bad outcome. One goroutine may want to log something, other want to set the adapter. The first one will realise that no adapter is set, the second one will set the adapter and then the first one will set the noop adapter. So eventually setting the adapter to specific implementation will not work.