Convert logger.Local from function to struct
At the moment logger.Local(adapter) is a constructor creating LocalLogger instance. This constructor does not do anything special. That's why it is not needed.

Also, other structures in yala (adapters) are initialized using struct literal. So using a function constructor is not very intuitive.

Also, global logger is named logger.Global. So local one should be named logger.Local.